Added self-rating to nomination form.
